Split utility and startup code into multiple translation units for better link-order matching
