+++ src/shared/util2.c
@@ -10,28 +10,20 @@ int openFile(const char *name, int mode);
 void fileClose(int handle);
 void decodePic(int handle, int segment);
 
-int openFileWrapper(const char *filename, int mode)
-{
-    return openFile(filename, mode);
-}
-
 void mystrcpy(char *dest, const char *source) {
     do {
     } while ((*dest++ = *source++) != '\0');
+}
+
+int openFileWrapper(const char *filename, int mode)
+{
+    return openFile(filename, mode);
 }
 
 void closeFileWrapper(int handle)
 {
     TRACE(("closeFileWrapper"));
     fileClose(handle);
-}
-
-void loadPic(const char *filename, int segment) {
-    int handle;
-    handle = openFileWrapper(filename, 0);
-    TRACE(("loadPic(): opened %s, loading into segment 0x%x", filename, segment));
-    decodePic(handle, segment);
-    closeFileWrapper(handle);
 }
 
 #ifdef BUGFIX
@@ -51,4 +43,12 @@ void openShowPic(char *name, int page, int garbage)
 #endif
     closeFileWrapper(fileHandle);
     TRACE(("openShowPic: file closed, returning"));
+}
+
+void loadPic(const char *filename, int segment) {
+    int handle;
+    handle = openFileWrapper(filename, 0);
+    TRACE(("loadPic(): opened %s, loading into segment 0x%x", filename, segment));
+    decodePic(handle, segment);
+    closeFileWrapper(handle);
 }
