+++ src/shared/util_c.c
@@ -1,0 +1,110 @@
+/*
+ * util_c.c - shared between start.exe and end.exe (split from util.c)
+ */
+
+#include "util.h"
+
+#include "../pointers.h"
+
+extern int far gfx_setFont(uint16 ch, uint16 font);
+
+void drawStringCentered(int16 *page, const char *str, int startx, int y, int endx) {
+    int width;
+    width = stringWidth(page, str);
+    drawStringAt(page, str, (endx - width) / 2 + startx, y);
+}
+
+int stringWidth(int16 *page, const char *str) {
+    int n;
+#ifdef BUGFIX
+    const char* l; // clang errors due to wrong type but char does change the binary
+#else
+    const uint8* l;
+#endif
+    int j;
+    l = str;
+    j = page[6];
+    n = 0;
+    while (*l != '\0') {
+        n += gfx_setFont(*(l++), j);
+    }
+    return n;
+}
+
+void my_ltoa(int32 value, char* buf) {
+    int8 i, k;
+#ifdef BUGFIX
+    char *p;
+#else
+    int8 *p;
+#endif
+    int8 n[6];
+    p = buf;
+    if (value < 0) {
+        value = -value;
+        *p = '-';
+        p++;
+    }
+    n[0] = value % 0xa;
+    value /= 0xa;
+    n[1] = value % 0xa;
+    value /= 0xa;
+    n[2] = value % 0xa;
+    value /= 0xa;
+    n[3] = value % 0xa;
+    value /= 0xa;
+    n[4] = value % 0xa;
+    value /= 0xa;
+    n[5] = value % 0xa;
+    i = 0;
+    for (k = 5; k > 0; k--) {
+        if (n[k] != 0) break;
+    }
+    do {
+        if (k == 2 && i == 1) {
+            *p = ',';
+            p++;
+        }
+        *p = n[k] + '0';
+        i = 1;
+        p++;
+    } while (--k >= 0);
+    *p = '\0';
+}
+
+void my_itoa(int value, char *buf) {
+    int8 n[6];
+    int8 i, k;
+    char *p;
+    p = buf;
+    if (value < 0) {
+        value = -value;
+        *p = 0x2d;
+        p++;
+    }
+    n[0] = value % 0xa;
+    value /= 0xa;
+    n[1] = value % 0xa;
+    value /= 0xa;
+    n[2] = value % 0xa;
+    value /= 0xa;
+    n[3] = value % 0xa;
+    value /= 0xa;
+    n[4] = value % 0xa;
+    value /= 0xa;
+    n[5] = value % 0xa;
+    i = 0;
+    for (k = 5; k > 0; k--) {
+        if (n[k] != 0) break;
+    }
+    do {
+        if (k == 2 && i == 1) {
+            *p = 0x2c;
+            p++;
+        }
+        *p = n[k] + 0x30;
+        i = 1;
+        p++;
+    } while (--k >= 0);
+    *p = 0;
+}
